Add --run-timeout option
Add an option for a timeout for the overall run. To support that, this
adds a new, stateful `Timeout` type that can be compared to see which
timeout is currently the most applicable.

This also allows the idle timeout to be longer than the maximum timeout
for the `poll()` call.

Fixes #10 — Overall timeout parameter.