Remove time and chrono dependencies
This depended on time and chrono indirectly via duration-str. It appears
that since rederr was not multi-threaded, it was not vulnerable to the
recent [vulnerability in time][] (chrono may have had the same
underlying problem).

This upgrades duration-str and disables all features on it, which
removes a bunch of indirect dependencies, including time and chrono.

[vulnerability in time]: https://github.com/rustsec/advisory-db/blob/df51dcdf972bebb7d2b1c4ae5861cc0f63dc60e6/crates/time/RUSTSEC-2020-0071.md