PR Checks: fix workflow to run after merging a PR
Previously the job would only run when opening or updating a PR. This
was unintentional, and a result of a misunderstanding of the
`$default-branch` “macro”, which is only used in workflow _templates_.
In actual workflows, it does nothing. This just changes to explicitly
using the branch name `main`.

It’s important that the job run after merge in order to update the
cache. The cache that’s saved when the PR is opened or updated is
[isolated to just that PR][isolation], whereas the cache on `main` can
be accessed by future PRs.

[isolation]: https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#restrictions-for-accessing-a-cache